fix(sca): Make narrowing conversions explicit
Document the intended loss of precision when storing sensor and parsed values in 32-bit fields. Match variadic format arguments to the unsigned integer type expected by %X while preserving the existing lower-32-bit output.

+++ src/SensorManager/Baro.cpp
@@ -51,7 +51,10 @@ void Baro::update_sensor(struct k_work *work) {
 	msg_baro.data.size = 2 * sizeof(float);
 	msg_baro.data.time = micros();
 
-	float data[2] = {bmp.temperature, bmp.pressure};
+	float data[2] = {
+		static_cast<float>(bmp.temperature),
+		static_cast<float>(bmp.pressure),
+	};
 
 	memcpy(msg_baro.data.data, data, 2 * sizeof(float));
